Moved validations
